misc: refactor ci, add osx-arm64 version, remove deps, bump miniz to 2.2.0
