--- .github/workflows/build-native.yml
@@ -1,61 +1,0 @@
-name: Build Native
-on: [push, pull_request]
-
-jobs:
-  build-native-win-x64:
-    runs-on: windows-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Build
-        run: |
-          cd build/libminiz
-          cmake -G "Visual Studio 16 2019" -A x64 -B build
-          cmake --build build --config Release
-      - uses: actions/upload-artifact@v2
-        with:
-          name: libminiz-win-x64
-          path: build/libminiz/build/lib/
-          
-  build-native-win-x86:
-    runs-on: windows-latest
-    steps:
-      - uses: actions/checkout@v2
-      - name: Build
-        run: |
-          cd build/libminiz
-          cmake -G "Visual Studio 16 2019" -A Win32 -B build
-          cmake --build build --config Release
-      - uses: actions/upload-artifact@v2
-        with:
-          name: libminiz-win-x86
-          path: build/libminiz/build/lib/
-        
-  build-native-osx-x64:
-    runs-on: macos-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: nuget/setup-nuget@v1
-      - name: Build
-        run: |
-          cd build/libminiz
-          cmake -B build
-          cmake --build build
-      - uses: actions/upload-artifact@v2
-        with:
-          name: libminiz-osx-x64
-          path: build/libminiz/build/lib/
-        
-  build-native-linux-x64:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: nuget/setup-nuget@v1
-      - name: Build
-        run: |
-          cd build/libminiz
-          cmake -B build
-          cmake --build build
-      - uses: actions/upload-artifact@v2
-        with:
-          name: libminiz-linux-x64
-          path: build/libminiz/build/lib/
