--- .github/workflows/nuget-native.yml
@@ -1,123 +1,0 @@
-name: Nuget Native
-on:
-  workflow_run:
-    workflows: [ "Build Native" ]
-    branches: [ master ]
-    types:
-      - completed
-
-env:
-  SECRET_PUBLISH: ${{ secrets.PUBLISH }}
-
-jobs:
-  nuget-native-win-x64:
-    runs-on: windows-latest
-    if: ${{ github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' }}
-    steps:
-      - uses: actions/checkout@v2
-      - uses: nuget/setup-nuget@v1
-      - name: Build
-        run: |
-          cd build/libminiz
-          cmake -G "Visual Studio 16 2019" -A x64 -B build
-          cmake --build build --config Release
-      - name: Pack
-        run: |
-          nuget pack build/native/NetMiniZ.Native.win-x64.nuspec -OutputDirectory nupkg
-      - name: Publish
-        if: ${{ env.SECRET_PUBLISH == 'true' }}
-        run: |
-          nuget push nupkg/*.nupkg -ApiKey ${{ secrets.NUGET_KEY }} -src https://api.nuget.org/v3/index.json
-        shell: bash
-      - uses: actions/upload-artifact@v2
-        with: 
-          name: NetMiniZ.Native.nuget
-          path: nupkg/
-          
-  nuget-native-win-x86:
-    runs-on: windows-latest
-    if: ${{ github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' }}
-    steps:
-      - uses: actions/checkout@v2
-      - uses: nuget/setup-nuget@v1
-      - name: Build
-        run: |
-          cd build/libminiz
-          cmake -G "Visual Studio 16 2019" -A Win32 -B build
-          cmake --build build --config Release
-      - name: Pack
-        run: |
-          nuget pack build/native/NetMiniZ.Native.win-x86.nuspec -OutputDirectory nupkg
-      - name: Publish
-        if: ${{ env.SECRET_PUBLISH == 'true' }}
-        run: |
-          nuget push nupkg/*.nupkg -ApiKey ${{ secrets.NUGET_KEY }} -src https://api.nuget.org/v3/index.json
-        shell: bash
-      - uses: actions/upload-artifact@v2
-        with: 
-          name: NetMiniZ.Native.nuget
-          path: nupkg/
-        
-  nuget-native-osx-x64:
-    runs-on: macos-latest
-    if: ${{ github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' }}
-    steps:
-      - uses: actions/checkout@v2
-      - uses: nuget/setup-nuget@v1
-      - name: Build
-        run: |
-          cd build/libminiz
-          cmake -B build
-          cmake --build build
-      - name: Pack
-        run: |
-          nuget pack build/native/NetMiniZ.Native.osx-x64.nuspec -OutputDirectory nupkg
-      - name: Publish
-        if: ${{ env.SECRET_PUBLISH == 'true' }}
-        run: |
-          nuget push nupkg/*.nupkg -ApiKey ${{ secrets.NUGET_KEY }} -src https://api.nuget.org/v3/index.json
-      - uses: actions/upload-artifact@v2
-        with: 
-          name: NetMiniZ.Native.nuget
-          path: nupkg/
-        
-  nuget-native-linux-x64:
-    runs-on: ubuntu-latest
-    if: ${{ github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' }}
-    steps:
-      - uses: actions/checkout@v2
-      - uses: nuget/setup-nuget@v1
-      - name: Build
-        run: |
-          cd build/libminiz
-          cmake -B build
-          cmake --build build
-      - name: Pack
-        run: |
-          nuget pack build/native/NetMiniZ.Native.linux-x64.nuspec -OutputDirectory nupkg
-      - name: Publish
-        if: ${{ env.SECRET_PUBLISH == 'true' }}
-        run: |
-          nuget push nupkg/*.nupkg -ApiKey ${{ secrets.NUGET_KEY }} -src https://api.nuget.org/v3/index.json
-      - uses: actions/upload-artifact@v2
-        with: 
-          name: NetMiniZ.Native.nuget
-          path: nupkg/
-
-  nuget-native-native:
-    runs-on: ubuntu-latest
-    if: ${{ github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' }}
-    steps:
-      - uses: actions/checkout@v2
-      - uses: nuget/setup-nuget@v1
-      - name: Pack
-        run: |
-          nuget pack build/native/NetMiniZ.Native.nuspec -OutputDirectory nupkg
-      - name: Publish
-        if: ${{ env.SECRET_PUBLISH == 'true' }}
-        run: |
-          nuget push nupkg/*.nupkg -ApiKey ${{ secrets.NUGET_KEY }} -src https://api.nuget.org/v3/index.json
-      - uses: actions/upload-artifact@v2
-        with:
-          name: NetMiniZ.Native.nuget
-          path: nupkg/
