--- .github/workflows/nuget.yml
@@ -1,32 +1,0 @@
-name: Nuget
-on:
-  workflow_run:
-    workflows: [ "Build" ]
-    branches: [ master ]
-    types:
-      - completed
-
-env:
-  SECRET_PUBLISH: ${{ secrets.PUBLISH }}
-
-jobs:
-  nuget:
-    runs-on: ubuntu-latest
-    if: ${{ github.event.workflow_run.conclusion == 'success' && github.event.workflow_run.event == 'push' }}
-    steps:
-      - uses: actions/checkout@v2
-      - name: Build
-        run: |
-          dotnet build -c Release /p:TargetOS=OSX
-          dotnet build -c Release /p:TargetOS=Linux
-      - name: Pack
-        run: |
-          dotnet pack -c Release /p:TargetOS=Windows -o nupkg
-      - name: Publish
-        if: ${{ env.SECRET_PUBLISH == 'true' }}
-        run: |
-          dotnet nuget push nupkg/*.nupkg -k ${{ secrets.NUGET_KEY }} -s https://api.nuget.org/v3/index.json
-      - uses: actions/upload-artifact@v2
-        with: 
-          name: NetMiniZ.nuget
-          path: nupkg/
