+++ native/libminiz/CMakeLists.txt
@@ -1,8 +1,8 @@
-cmake_minimum_required(VERSION 3.14)
+cmake_minimum_required(VERSION 3.21)
 
 include(FetchContent)
 
-set(PROJECT_VERSION "2.1.0")
+set(PROJECT_VERSION "2.2.0")
 
 project(miniz VERSION ${PROJECT_VERSION})
 
@@ -22,18 +22,17 @@ target_include_directories(miniz PUBLIC
 )
 
 if(MSVC)
-   set_target_properties(miniz 
-      PROPERTIES VERSION ${PROJECT_VERSION}
-      RUNTIME_OUTPUT_DIRECTORY_RELEASE "${CMAKE_BINARY_DIR}/lib"
+   set_target_properties(miniz PROPERTIES 
+      VERSION ${PROJECT_VERSION}
+      RUNTIME_OUTPUT_DIRECTORY_RELEASE "${CMAKE_BINARY_DIR}"
       RUNTIME_OUTPUT_NAME "libminiz-${PROJECT_VERSION}"
    )
 else()
-   set_target_properties(miniz 
-      PROPERTIES VERSION ${PROJECT_VERSION}
-      LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib"
+   set_target_properties(miniz PROPERTIES 
+      VERSION ${PROJECT_VERSION}
+      LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}"
    )
 endif()
 
 file(COPY ${minizsrc_SOURCE_DIR}/LICENSE ${minizsrc_SOURCE_DIR}/ChangeLog.md
-     DESTINATION ${CMAKE_BINARY_DIR}/lib)
-
+     DESTINATION ${CMAKE_BINARY_DIR})
