+++ native/nuget/NetMiniZ.Native.osx-arm64.nuspec
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<package xmlns="http://schemas.microsoft.com/packaging/2013/01/nuspec.xsd">
+  <metadata>
+    <!-- package -->
+    <id>NetMiniZ.Native.osx-arm64</id>
+    <title>NetMiniZ - Native binaries for osx-arm64</title>
+    <version>__VERSION__</version>
+    <description>This package complements the NetMiniZ package and contains native binaries of libminiz for osx-arm64</description>
+    <summary>Native binaries of libminiz for osx-arm64</summary>
+    <projectUrl>https://github.com/jsm174/net-miniz</projectUrl>
+    <repository type="git" url="https://github.com/jsm174/net-miniz" />
+    <tags>libminiz binaries</tags>
+    <!-- legal -->
+    <license type="expression">MIT</license>
+    <authors>Jason Millard</authors>
+    <owners>NetMiniZ development team and contributors</owners>
+    <requireLicenseAcceptance>false</requireLicenseAcceptance>
+    <copyright>NetMiniZ development team and contributors</copyright>
+    <dependencies>
+      <group targetFramework=".NETStandard2.0" />
+    </dependencies>
+  </metadata>
+  <files>
+    <!-- The build bits -->
+    <file src="targets\NetMiniZ.Native.osx-arm64.targets" target="build\netstandard2.0" />
+    <!-- Include libminiz osx-arm64 binaries -->
+    <file src="libminiz-osx-arm64\libminiz.*.dylib" target="runtimes\osx-arm64\native" />
+    <!-- Include the license -->
+    <file src="..\..\LICENSE.txt" />
+    <!-- A dummy reference which prevents NuGet from adding any compilation references when this package is imported -->
+    <file src="_._" target="lib\netstandard2.0" />
+  </files>
+</package>
