+++ native/nuget/NetMiniZ.Native.win-x64.nuspec
@@ -4,7 +4,7 @@
     <!-- package -->
     <id>NetMiniZ.Native.win-x64</id>
     <title>NetMiniZ - Native binaries for win-x64</title>
-    <version>1.2.2</version>
+    <version>__VERSION__</version>
     <description>This package complements the NetMiniZ package and contains native binaries of libminiz for win-x64</description>
     <summary>Native binaries of libminiz for win-x64</summary>
     <projectUrl>https://github.com/jsm174/net-miniz</projectUrl>
@@ -13,19 +13,20 @@
     <!-- legal -->
     <license type="expression">MIT</license>
     <authors>Jason Millard</authors>
-    <owners>Jason Millard</owners>
+    <owners>NetMiniZ development team and contributors</owners>
     <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <copyright>Jason Millard</copyright>
+    <copyright>NetMiniZ development team and contributors</copyright>
+    <dependencies>
+      <group targetFramework=".NETStandard2.0" />
+    </dependencies>
   </metadata>
   <files>
     <!-- The build bits -->
     <file src="targets\NetMiniZ.Native.win-x64.targets" target="build\netstandard2.0" />
     <!-- Include libminiz win-x64 binaries -->
-    <file src="..\libminiz\build\lib\libminiz-*.dll" target="runtimes\win-x64\native" />
+    <file src="libminiz-win-x64\libminiz-*.dll" target="runtimes\win-x64\native" />
     <!-- Include the license -->
-    <file src="..\libminiz\build\lib\LICENSE" />
-    <!-- Include the change log -->
-    <file src="..\libminiz\build\lib\ChangeLog.md" />
+    <file src="..\..\LICENSE.txt" />
     <!-- A dummy reference which prevents NuGet from adding any compilation references when this package is imported -->
     <file src="_._" target="lib\netstandard2.0" />
   </files>
