+++ native/nuget/targets/NetMiniZ.Native.osx-arm64.targets
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup Condition="'$(MSBuildRuntimeType)' == 'Mono'">
+    <None Include="$(MSBuildThisFileDirectory)..\..\runtimes\osx-arm64\native\*">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+</Project>
