+++ native/nuget/targets/NetMiniZ.Native.osx-x64.targets
