+++ native/nuget/targets/NetMiniZ.Native.win-x64.targets
