+++ src/NetMiniZ/BaseException.cs
@@ -1,16 +1,16 @@
-﻿namespace NetMiniZ
+﻿using System;
+
+namespace NetMiniZ
 {
-    using System;
+	public abstract class BaseException : Exception
+	{
+		public string ComponentName { get; }
+		public int Status { get; }
 
-    public abstract class BaseException : Exception
-    {
-        public string ComponentName { get; }
-        public int Status { get; }
-
-        public BaseException(string componentName, int status)
-        {
-            ComponentName = componentName;
-            Status = status;
-        }
-    }
+		public BaseException(string componentName, int status)
+		{
+			ComponentName = componentName;
+			Status = status;
+		}
+	}
 }
