+++ src/NetMiniZ/CompressException.cs
@@ -1,17 +1,17 @@
 ﻿namespace NetMiniZ
 {
-    public class CompressException : BaseException
-    {
-        public CompressException(string componentName, int status)
-            : base(componentName, status)
-        { }
+	public class CompressException : BaseException
+	{
+		public CompressException(string componentName, int status) : base(componentName, status)
+		{
+		}
 
-        public override string Message
-        {
-            get
-            {
-                return string.Format("Compression routine {0} failed with error code {1}.", ComponentName, Status);
-            }
-        }
-    }
+		public override string Message
+		{
+			get
+			{
+				return string.Format("Compression routine {0} failed with error code {1}.", ComponentName, Status);
+			}
+		}
+	}
 }
