+++ src/NetMiniZ/DecompressException.cs
@@ -1,17 +1,17 @@
 ﻿namespace NetMiniZ
 {
-    public class DecompressException : BaseException
-    {
-        public DecompressException(string componentName, int status)
-            : base(componentName, status)
-        { }
+	public class DecompressException : BaseException
+	{
+		public DecompressException(string componentName, int status) : base(componentName, status)
+		{
+		}
 
-        public override string Message
-        {
-            get
-            {
-                return string.Format("Decompression routine {0} failed with error code {1}.", ComponentName, Status);
-            }
-        }
-    }
+		public override string Message
+		{
+			get
+			{
+				return string.Format("Decompression routine {0} failed with error code {1}.", ComponentName, Status);
+			}
+		}
+	}
 }
