+++ src/NetMiniZ/Interop/Libraries.Linux.cs
@@ -1,7 +1,7 @@
 namespace NetMiniZ.Interop
 {
-    internal static partial class Libraries
-    {
-        internal const string MiniZ = "libminiz.so.2.1.0";
-    }
+	internal static partial class Libraries
+	{
+		internal const string MiniZ = "libminiz.so.2.2.0";
+	}
 }
