+++ src/NetMiniZ/Interop/Libraries.OSX.cs
@@ -1,7 +1,7 @@
 namespace NetMiniZ.Interop
 {
-    internal static partial class Libraries
-    {
-        internal const string MiniZ = "libminiz.2.1.0.dylib";
-    }
+	internal static partial class Libraries
+	{
+		internal const string MiniZ = "libminiz.2.2.0.dylib";
+	}
 }
