+++ src/NetMiniZ/Interop/Libraries.Windows.cs
@@ -1,7 +1,7 @@
 namespace NetMiniZ.Interop
 {
-    internal static partial class Libraries
-    {
-        internal const string MiniZ = "libminiz-2.1.0.dll";
-    }
+	internal static partial class Libraries
+	{
+		internal const string MiniZ = "libminiz-2.2.0.dll";
+	}
 }
